Revert "plutus-playground: remove 2 tests"
Bring back the previously removed tests. The removal was because ghc
with relevant packages wasn't provided to the derivation that creates
the purescript bridge code. In doing so (and also adding it to the dev
script that creates the generated dir) we can easily bring back the
tests and the creation of the relevant data from PSGenerator in the
plutus-playground-server.